Don't timeout unless specifically set
By default, the spinner will timeout at 5 seconds, regardless of when `hide()` is called. This leads to the less-than-optimal workaround of adding an arbitrarily long value for the timeout attribute, e.g. `<ng4-loading-spinner [timeout]="99000"></ng4-loading-spinner>` This proposal makes it so the default is zero, and if zero, will not time out and hide the spinner before `hide()` is called.